Reorganized frontend structure to reduce the amount of unwanted files showing up
