+++ frontend/src/components/About/About.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './About.css'; 
+import '../About/About.css'; 
 
 function About() {
   return (
