+++ frontend/src/components/Calendar/Sidebar.css
